Enlarging chunk size to 1M bytes; Change type of elapse from float64 to int64.
original_cl=101241
submit_cl=101242

+++ src/jingoal/dfs/metadata/event.go
@@ -70,7 +70,7 @@ type Event struct {
 	Domain      int64         `bson:"domain"`              // domain
 	EType       EventType     `bson:"eType"`               // event type (dfs 2.0)
 	Fid         string        `bson:"fid"`                 // fid
-	Elapse      float64       `bson:"elapse,omitempty"`    // elapse in millisecond
+	Elapse      int64         `bson:"elapse,omitempty"`    // elapse in nanosecond
 }
 
 type EventOp struct {
